+++ src/components/header/header_links/header_links.tsx
@@ -137,6 +137,7 @@ export const EuiHeaderLinks: FunctionComponent<EuiHeaderLinksProps> = ({
               anchorPosition="downRight"
               closePopover={closeMenu}
               panelPaddingSize="none"
+              repositionOnScroll
               {...popoverProps}
             >
               <div